initial

+++ src/main/java/com/gunitha/site_management_system_java_backend/model/update/SiteInfoUpdate.java
@@ -20,9 +20,9 @@ public class SiteInfoUpdate {
 
     private String number;
 
-    private Integer width;
+    private Double width;
 
-    private Integer length;
+    private Double length;
 
     private String description;
